# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inazuma API",
        "description": "API for Inazuma Eleven",
        "contact": {
            "email": "gomez.sergio.smr@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/v1/emblems": {
            "get": {
                "tags": [
                    "Emblems"
                ],
                "summary": "List all emblems",
                "operationId": "bcd941e46969e46e66954b7c7d8927ca",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "name[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Raimon"
                        }
                    },
                    {
                        "name": "version[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "ie1"
                        }
                    },
                    {
                        "name": "version[ne]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "ie1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmblemCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Emblems"
                ],
                "summary": "Create new emblem",
                "operationId": "8b189e51accf2123f5dd2b1f05a292f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmblemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Emblem created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmblemResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/emblems/{id}": {
            "get": {
                "tags": [
                    "Emblems"
                ],
                "summary": "Get specific emblem",
                "operationId": "cc89073512aaf4d258fe3379f6088b20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Emblem details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmblemResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Emblem not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Emblems"
                ],
                "summary": "Update an emblem",
                "operationId": "5445e667b7b4772d14c960d1d41ad280",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmblemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Emblem updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmblemResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Emblems"
                ],
                "summary": "Delete emblem",
                "operationId": "d71ec0a636e51bb7276932e576f4d305",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Emblem deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Emblem deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/formations": {
            "get": {
                "tags": [
                    "Formations"
                ],
                "summary": "List all formations",
                "operationId": "c0436c2b13529fc03997bccef27302a8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "name[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Diamante"
                        }
                    },
                    {
                        "name": "layout[eq]",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "4-3-3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormationCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Formations"
                ],
                "summary": "Create new formation",
                "operationId": "fd590d2d4ef87ac82d3380be22cd5fe3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Formation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/formations/{id}": {
            "get": {
                "tags": [
                    "Formations"
                ],
                "summary": "Get specific formation",
                "operationId": "e8e2af8f9d62862d8aa6a0501f13037a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Formation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Formations"
                ],
                "summary": "Update a formation",
                "operationId": "b226c1d85416ff7b6810d2f4753848da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Formation updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Formations"
                ],
                "summary": "Delete formation",
                "operationId": "897701a5401c554d1a839c2510d9bd58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Formation deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Formation deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmblemRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Emblem name",
                        "type": "string",
                        "example": "Raimon"
                    },
                    "image": {
                        "description": "URL to emblem image",
                        "type": "string",
                        "example": "https://static.wikia.nocookie.net/inazuma/images/e/e1/Raimon_FF_Emblema.png/revision/latest?cb=20210620190405&path-prefix=es"
                    },
                    "version": {
                        "description": "Game version",
                        "type": "string",
                        "example": "ie1"
                    }
                },
                "type": "object"
            },
            "EmblemResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Emblem"
                    }
                },
                "type": "object"
            },
            "Emblem": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Raimon"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://static.wikia.nocookie.net/inazuma/images/e/e1/Raimon_FF_Emblema.png/revision/latest?cb=20210620190405&path-prefix=es"
                    },
                    "version": {
                        "type": "string",
                        "example": "ie1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EmblemCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Emblem"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/emblems?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/emblems?page=3"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/emblems?page=2"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 35
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "FormationRequest": {
                "required": [
                    "name",
                    "layout"
                ],
                "properties": {
                    "name": {
                        "description": "Formation name",
                        "type": "string",
                        "example": "Diamante"
                    },
                    "layout": {
                        "description": "Formation layout pattern",
                        "type": "string",
                        "example": "4-3-3"
                    }
                },
                "type": "object"
            },
            "FormationResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Formation"
                    }
                },
                "type": "object"
            },
            "Formation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Diamante"
                    },
                    "layout": {
                        "type": "string",
                        "example": "4-3-3"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "FormationCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Formation"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/formations?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/formations?page=3"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/formations?page=2"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 35
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}